ft-#16: complete search-daycares page functionality and filter
