+++ src/components/Filter/type.ts
@@ -1,13 +1,21 @@
-import { TChildCare } from "../../pages/Home/type"
 import { TFiltered } from "../../pages/Result/type"
 
 type TFilter = {
     setFilter: React.Dispatch<React.SetStateAction<string>>
-    setIsFilterClicked: React.Dispatch<React.SetStateAction<boolean>>,
-    setIsFilterLoading: React.Dispatch<React.SetStateAction<boolean>>,
-    data: TChildCare[] | undefined,
+    setIsFilterClicked: React.Dispatch<React.SetStateAction<boolean>>
     setFiltered: React.Dispatch<React.SetStateAction<TFiltered>>
 
 }
 
-export  type {TFilter}
+
+type TFilterPayload = {
+    location: string,
+    maxp: string,
+    minp: string
+}
+
+type TDefault = {
+    location: string
+}
+
+export  type {TFilter, TFilterPayload, TDefault}
