+++ src/pages/Result/hooks/useSearch.ts
@@ -2,9 +2,11 @@ import { useEffect, useState } from "react"
 import { useSearchParams } from "react-router-dom"
 import useSearchDayCares from "./useSearchDaycare"
 import useLastSearch from "../../Search/hooks/useLastSearch"
+import { TFiltered } from "../type"
+import { TChildCare } from "../../Home/type"
 
 const useSearch  = ()=>{
-
+    const [filtered, setFiltered] = useState<TFiltered>({location:"", minp:0, maxp: 600})
     let [search, setSearchParams] = useSearchParams()
     let {data, setTerm, isLoading, setData} = useSearchDayCares()
     const [input, setInput] = useState<string>()
@@ -22,6 +24,7 @@ const useSearch  = ()=>{
         e.preventDefault()
         if(input){
             setSearchParams({term: input})
+            setFiltered({location: '', minp:0, maxp: 600})
             handleLastSearch(input)
         }
     }
@@ -31,7 +34,18 @@ const useSearch  = ()=>{
         setInput(e.target.value)
     }
 
-    return {onSearchTermChanged, handleInput, filter,setFilter,data,input,isLoading, isFilterClicked,setIsFilterClicked,setData}
+    let filterData = (data: TChildCare[])=>{
+        let  filt =  data.filter((item)=> item.amount <= filtered.maxp)
+        .filter((item)=> item.amount >= filtered.minp)
+
+        if(filtered.location !== ""){
+           return filt.filter((item)=> item.exactLocation.toLowerCase() === filtered.location.toLowerCase())
+        }
+        return filt
+    }
+
+
+    return {onSearchTermChanged, handleInput, filter,setFilter,data,input,isLoading, isFilterClicked,setIsFilterClicked,setFiltered, filterData}
 }
 
 export default useSearch
