+++ src/pages/Result/index.tsx
@@ -9,23 +9,10 @@ import useSearch from "./hooks/useSearch"
 import Modal from "../../components/Modal"
 import Filter from "../../components/Filter"
 import ResultWrapper from "./ResultWrapper"
-import {useState } from "react"
-import { TChildCare } from "../Home/type"
-import { TFiltered } from "./type"
-
 
 const Result = ()=>{
 
-    const {data,filter,handleInput,onSearchTermChanged,setFilter, input,isLoading, setIsFilterClicked,isFilterClicked,setData} = useSearch()
-    const [isFilterLoading, setIsFilterLoading] = useState<boolean>(false)
-    const [filtered, setFiltered] = useState<TFiltered>({location:"", minp:0, maxp: 600})
-
-    let filterData = (data: TChildCare[])=>{
-        let  filt =  data.filter((item)=> item.amount <= filtered.maxp)
-        .filter((item)=> item.amount >= filtered.minp)
-        .filter((item)=> item.exactLocation.includes(filtered.location))
-        return filt
-    }
+    const {data,filter,handleInput,onSearchTermChanged,setFilter, input,isLoading, setIsFilterClicked,isFilterClicked, setFiltered, filterData} = useSearch()
 
     return (
         <ResultWrapper>
@@ -58,10 +45,10 @@ const Result = ()=>{
 
         {data && data?.length < 1 && <p className="no-data">sorry we don't have a daycare for the specified search term, try search  with different keywords</p>}
 
-        {(isLoading || isFilterLoading ) &&  <p className="loader"></p>}
+        {isLoading  &&  <p className="loader"></p>}
 
         {isFilterClicked && <Modal ismodalopen={JSON.stringify(isLoading)}>
-            <Filter setFiltered={setFiltered}  data={data}  setIsFilterLoading={setIsFilterLoading} setFilter={setFilter} setIsFilterClicked={setIsFilterClicked} />
+            <Filter setFiltered={setFiltered}  setFilter={setFilter} setIsFilterClicked={setIsFilterClicked} />
             </Modal>}
         </div>
